Download and install deb only if it's not installed already
Problems with original approach:
1) apt wuth deb always downloads the file. Afterwards, it checks if the
package is installed already and takes the appropriate action. This is
slow.

I could not find a way to ensure the package is not downloaded first.

2) Let's add a `when`f. I tried to put the check and register to a
variable. The problem is that ansible does not support dynamic variable
names, so could not do sth like
`register: '{{item}}'_installed`
if I understand correctly this means I must define a static variable
name per package I want to install.

So I ended up exploding this `with_items` to all these lines

Maybe there is a better way to do this?